requirements.txt
# Core LangChain dependencies
langchain>=0.3.27,<0.4
langgraph>=0.6.8,<0.7

# LLM providers
langchain-anthropic>=0.3.21,<0.4
langchain-openai>=0.3.0,<0.4

# Environment and utilities
python-dotenv>=1.0.0

# Search providers
duckduckgo-search>=6.0.0
requests>=2.31.0

# Web interface
gradio>=4.0.0

# Type hints for Python <3.10
typing-extensions>=4.5.0
